Trait io_providers::env::Provider

pub trait Provider {
    fn args(&self) -> Vec<String>;
    fn current_dir(&self) -> Result<PathBuf>;
    fn set_current_dir(&mut self, path: &Path) -> Result<()>;
}

Provides access to environment data, such as working directory and environment variables.

This trait acts more-or-less as a drop-in replacement for std::env functions. The main differences are:

env::Provider::args() returns Vec<String>, and not an iterator like std::env::args does.

All of the path parameters on env::Provider methods require &Path as opposed to AsRef<Path>. In order to allow trait objects, env::Provider must have object safety, and this requires non-generic methods (among other things). As such, we have to put up with the less ergonomic &Path.

Required Methods

fn args(&self) -> Vec<String>

Returns the arguments which this program was started with (normally passed via the command line).

See std::env::current_dir for more information.

fn current_dir(&self) -> Result<PathBuf>

Returns the current working directory as a PathBuf.

See std::env::current_dir for more information.

fn set_current_dir(&mut self, path: &Path) -> Result<()>

Changes the current working directory to the specified path, returning whether the change was completed successfully or not.

See std::env::set_current_dir for more information.